Introducing Oxygen icons
The desktop flavours of the AlarmNotfications application need some
icons for their GUI, especially the context menu of the tray icon.

These icons are taken from the Oxygen Project within KDE
(https://techbase.kde.org/Projects/Oxygen). The icons are licensed
LGPL v3 so we can use them.

To fulfill the condition of conveying the corresponding "source" to a
binary file (in this case the PNG icon files), the vector graphic files
where the icons originally originated from are also included.