Flashlight driver class
An alarm can be anounced in the laboratory by switching on a red
flashlight. At Ruhr-Universität Bochum, we have a "KMTronic USB One
Relay", so this is the only one supported by this class.

The API to the outside is rather easy: There are just the two static
methods switchOn() and switchOff(), so adapting the class to other
hardware should be easy. If the other hardware is also controlled via a
serial interface, it might be enough to change the createCommand()
method.