Class for sending alarm notifcation e-mails
The class EMailSender exists in two different implementations:

The "real" implementation in emailsender.cpp links to the QtSmtpClient
library and can assemble and transmit e-mail messages.

The dummy implementation of the same class is located in
emailsender_dummy.cpp and does nothing. This one will be linked into the
desktop flavours of AlarmNotifications to satisfy the dependencies of
classes common to both the desktop and daemon flavours.

Both implementations share the same header file.