Beedo engine: Opto-acoustic alarms
The beedo engine is used in the beamtime flavours of AlarmNotifications.
It plays a video in the top left corner of the screen that cannot be
"clicked away".

Please note that the Beedo engine will not be activated unless a video
file named "beedo.ogv" is present in the source code tree. As such a
file is site-specific, it is not included in the public repository.