Configuration utility for AlarmNotifications
The an-config binary is a separate utility that provides a GUI for
configuring all the flavours of AlarmNotifications.

The configuration file is a plain text file in INI style, so
configuration can also be done with any text editor. The available
configuration options are documented in README.md and on the start page
of the Doxygen documentation.