AlarmNotifications daemon class
The Daemon class provides the daemon flavour of AlarmNotifications that
is intended to run on a central lab computer or a server. There can be
any number of daemons on the network (including 0, the desktop flavours
don't need it).

The daemon flavour supports all three kinds of notifications: In case of
an alarm, it can enable a flashlight in the laboratory, show a
notification on the lab desktop and send out e-mails in case there is
nobody to fix the issue.