Desktop flavours of the AlarmNotifications
There are two implementations that have a common abstrace base class:

- DesktopWidgetQt uses the legacy QSystemTrayIcon which is supported on
  older systems but lacks some features compared to modern ones.
- DesktopWidgetKde4 uses the KStatusNotifierItem API which is based on a
  recent specification among free desktop environments. The tray icon
  can blink or hide itself and show more details in the tooltip.

From a functional point of view, both implementations are identical. The
DesktopWidgetKde4 implementations needs at least KDE SC Platform 4.4 to work.

Both implementations support the Beedo engine - it is activated if the
compiler flag -DBEEDO is set. The an-desktop and an-desktop-beamtime
binaries are therefore made from the same source code, just compiled
with different parameters. The same applies to the KDE4 flavours.